@@ -1,6 +1,6 @@
 module gitlab.com/postgres-ai/database-lab/v2
 
-go 1.15
+go 1.16
 
 require (
 	github.com/AlekSi/pointer v1.1.0
